requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
requests>=2.31,<3.0
scikit-learn>=1.3,<2.0
numpy>=1.24,<3.0
scipy>=1.10,<2.0
joblib>=1.2,<2.0
reportlab>=4.0,<5.0
pillow>=10.0,<11.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
